move depreciated functions to new file
